Added option to supply callback function when creating cluster. If a callback is provided, whenever a job's results are available, dispy will call supplied callback with that job as argument.
A computation may send provisional results (see earlier commit) back
to the client multiple times. By default each time the client retrieves
a job's result with 'job()', the event is cleared so that next 'job()'
will block until next (provisional) results are available. Now, it is
possible to call 'job(clear=False)' to leave the event in set state and
further calls to get results (with 'job(clear=False)') will not block, but
then the client will have to keep track of which jobs have new results.
This is getting a bit tricky even to explain; may be this feature needs to be
reworked or even dropped!