Fixed SSL setup.
When AsyncSocket is used, appropriate SSL certificates must be
passed. In the last few releases, this configuration has been
missing in some cases, because of which SSL was not being used
(correctly). This patch fixes those issues.

In dispyscheduler _job must be in use until the results are sent
to the client successfully. Otherwise, garbage oollector may reuse
same object for a new _job; this will cause 'id' to be same as the
id for another job that is still pending, causing duplicate IDs.
This has not been throughly tested to know if the fix is proper.