Recover from crashes is supported / improved.
Until now recover from client crashes is supported only if the
option is enabled. Now the client stores information needed for
recovery in persistence so that dispy.recover_jobs function can
be used to get the results of jobs that were scheduled at the
time of crash can be retrieved. dispynode now also stores the
information about computations so if the node crashes, any jobs
that were done but couldn't be sent before crash can be resent.

'status' function of a cluster can be used to retrieve status of
all the nodes executing jobs for that cluster. This information
is list of instances of DispyNode and number of jobs pending.
This feature is inspired from a patch from Dariusz Dwornikowski.
'stats' function is supported for backward comptability, but
'print_status' should be preferred.

Couple of users reported that nodes are not sending finished
jobs in some cases. I have not been able to reproduce / understand
this issue. My guess is that node may not able to connect to client
when jobs are terminated in some cases, so now the node tries to
send any pending results (saved in files) whenever it can send
other results when they are finished. Also poll_interval is set
to 10 minutes by default. This commit breaks poll_interval in other
ways, so that needs to be fixed before 4.3 release.