Polling for job results supported.
If nodes (or dispyscheduler) can't send job results to clien
(e.g., because client is behind firewall and port forwarding is
not possible), 'poll_interval' parameter can be used to poll for
job results at specified interval. This is inefficient, but if there
is no other way to use the compute resources, then this can be
used.