Revert patch that uses private global environment.
Using private global variables seemed like a nice idea, but many
things break with it (e.g., dispy_provisional_result / dispy_send_file)
are not available, and even if they are added to the environment,
other variables used in those functions are not available etc. If
it can be reintroduced, it needs more time / testing.