Added support for httpd for dispyscheduler.
This is (initial) support for monitoring and managing clusters / jobs
for dispyscheduler (used for SharedJobCluster clients).

Job IDs shown in 'Node' are different from those used in clients,
as dispyscheduler is not aware of what IDs user uses at the client.
The arguments / keyword arguments are not shown either, as these
come as pickled objects that are sent to nodes and dispyscheduler
doesn't load any code sent by clients so the objects can't be
unpickled. If correct values for these are needed, httpd can be
used at the client as well (i.e., another httpd can be run at client
while a httpd is running at dispyscheduler - on different ports if
both client and dispyscheduler are on same machine).