Fix dispyscheduler.py importing httpd.
dispy.py can also be used as a program, so it is installed as a
binary in the same directory where dispyscheduler.py is istalled.
Because of this, 'import dispy' imports the script where 'httpd'
module is not available - it is installed a sys.path directory.
The (not so clean) fix is to search sys.path where httpd module
is available and prepend it to sys.path till dispy is loaded.
Hopefully someone else has a better solution for this.