Initialize 'dest_path_prefix' with absolute path.
dispynode changes working directory to computation's directory.
With this, if 'dest_path_prefix' is given by as relative path,
the path used for saving files will be relative path wrt to
current working directory, instead of what the user specified.
So initialize 'dest_path_prefix' with the absolute path of what
the user specified.

This fixes bug #13 and solution is as suggested by the person
that filed the bug.