Maintain file use counts in cluster itself.
Files being transferred may already be present (e.g., if 'dest_path'
is used) at the scheduler and such files should not be removed when
computation is done. dispyscheduler uses a mapping to maintain file
use counts so that if use count drops to zero, file can be removed.
Maintain this mapping within cluster itself, instead of global,
similar to dispynode.