Use 51347 as port with SharedJobCluster.
Use port 51347 for client when used with SharedJobCluster. Earlier
a random (available) port is used, which is convenient if dispyscheduler
(which uses port 51347) runs on same computer. However, it is less
common for both to be running on same computer, and more likely
using random port confuses users that use firewall. Now port 51347
is used in the client as well, and if dispyscheduler runs on same
computer, then either client or dispyscheduler will fail so the issue
is at least more obvious.

If a node changes CPUs and is not used for a computation, setup
the node so it can be used.