Dynamically changing CPUs in dispynode fully supported.
With "cpus" command, the number of CPUs used for dispy clients can
be changed anytime. If the node is being used at the time, then
the node can inform the scheduler about the change. However, if
the client initially found a node with a certain number of CPUs
during discovery, this information may be wrong at the time a new
computation is submitted. dispynode now sends the currently
available processors when a computation is submitted. (Any changes
to CPUs later will be sent to the scheduler.)

If submitting a job fails, discard it from scheduling again;
otherwise the scheduler may end up submitting it again and again,
possibly causing livelock.