Setup node with coroutines.
As mentioned in previous commit, user's "setup" function is
called with no timeout. This can potentially hang client program
if setup of a node fails (but other nodes execute jobs and computation
is done). This patch fixes the problem by using coroutines (instead
of 'yield' to from 'add_node') and making 'setup_node' coroutine
a daemon (so scheduler will not block if otherwise computation is
finished).