Support keyword argument when computation is program.
'dispy_job_depends' keyword parameter is used to send job specific
dependencies (files). If computation is program, job creation now
refuses keyword parameters, thus job dependencies can't be used.
The fix (for now) is not to refuse keyword parameters. To be strict,
only 'dispy_job_depends' can be allowed, but checking can be
expensive and most of the time not necessary.