Fixed dispynode to work with Python 3 under Windows.
It seems Python 3 doesn't start multiprocessing.Process under
Windows if main program reads standard input (either with 'input'
function, or sys.stdin.readline function). To have the process
started, 'Enter' must be given twice - multiprocessing.Process
seems blocked on reading two lines of input. Until this is fixed,
disable reading input under Windows with Python 3 (so dispynode
works as if it is a daemon, as it used to prior to 4.6.3 release).