Added support for informing node availability status.
Node availability status (CPU, memory and disk space) is now sent
with pulse messages if 'psutil' module is installed. This information
is given to NodeAllocate's allocate method, for example, to filter available
nodes. This information is also sent to cluster_status callback, which
is used to update node's status in httpd as well (so it can be
monitored in web browser).