Call 'add_node' as regular function, not generator.
'add_node' is regular function, not generator. Calling it with
'yield' as generator with open sockets will leak socket as the
function does 'raise StopIteration' (it used to be generator before
but was later changed function, without changing 'raise' to 'return').

To prevent sockets from leaking, 'tcp_task' now closes connection
sockets as soon as they are not needed.

Thanks to Stelios Kyiracou for pointing this issue.