Fixed race condition with dispyscheduler.
When SharedJobCluster is used, if the jobs take very little time
to execute, there is a chance that the job finishes before client
gets ID for submitting a job, in which case job reply is ignored.
The fix is for client to first get the ID and put it in its pending
jobs and send ack to scheduler; only then scheduler submits job for
execution.