Fixed job termination.
When a running job is terminated (killed), the status must be set
to Terminated and result to None. Until now job process is killed
but status and result may not be set accordingly, causing client
program to not realize which jobs were terminated. This is now
fixed; job termination task and job process set the status and
result attributes appropriately both for computation functions
and computation programs.