Use default value as 'None' for certificate files.
SSL certificate file names must be either valid paths or 'None'.
Using empty string (default value as of now) will break SSL if
keyfile is empty (if both certificates are in same file, they
must be in 'certfile' in which case keyfile is left with default
value of '' which is not valid path).