Example for setup/cleanup fixed
When files are saved to the node, they stored under the working dir + computation_id path. If on the client files are defined with absolute paths they can't be loaded on the node.
Then setup function is failed some of global variables can be undefined. Cleanup function shouldn't rasise the NameError in case of deletion undefined global variables.