Prevent changes to os.environ.
'setup' and 'cleanup' execute in main dispynode thread. These
functions can change os.environ that may affect other computations,
so maintain original state and restore after computation is closed,
or if node is shared (with SharedJobCluster clients).

A better solution is to run part of dispynode in child process of
main so any such changes are automatically discarded when computation
is done (as done with dispycosnode). This will be done in the next
(or a future) version.