Throw exception if 'inet_pton' is not available.
Sockets under Windows may not have 'inet_pton' method. Pycos loads
'win_inet_pton' module (that adds this method to sockets), but
doesn't issue warning if it is not available (as this method is
needed only for IPv6). If dispy uses IPv6 and 'inet_pton' is not
available, throw exception with appropriate message.