Fixed recent dispynode changes to support Windows.
With Windows, setup process function must be global function
(instead of class function in _DispyNode). Currently with Windows,
global variables that can be initialized in setup are supposed to
be native types (i.e., they can't be instances of classes even
if they are sent with depends), so creating jobs is efficient.
Perhaps this is too limiting (and performance overhead with
sending job arguments over pipe to setup process)?