Many updates to dispynode computation model.
Generalized using a sub process to run "setup" / "cleanup"
functions. Now, when a computation uses these features, a sub
process is used to run the functions. If "setup" returns 0, the
sub process is used to start jobs only if any global variables
are defined in "setup" function and if no global variables are
defined, the main process creates jobs (so there is no overhead
of passing job args to sub process). If "setup" returns 1, then
it is assumed that even if global variables are defined in it,
the main process can import those variables and then start jobs
directly from main process (to avoid overhead), but this will
only work if global variables are of native Python types. If
"setup" returns 2, sub process is always used to start jobs,
even if no global variables are defined in "setup".

Improved process termination when client cancels jobs and when "clean"
option is used to terminate processes in prior run.