Don't throw away abandoned jobs (fix for #142).
If a job is abandoned (due to node becoming zombie), don't throw
it away; it is likely node may come alive while client is still
running. Instead, put them aside, inform callback about the status
(as Abandoned) and if reply for any such job is received, finish
it as normal.