Initialize 'id' field of jobs with unique ID.
Until now 'id' field of jobs is initialized to 'None'. User can
reset it to appropriate value. However, job may fail before user
has a chance to reset which may make it difficult to distinguish
failed job. Now the 'id' field is initialized to unique value
(that remains unique until job is done) and user can use that
value or reset it as appropriate.