Improved termination of job processes.
Job processes may not terminate / may become zombies, etc., so
these conditions should be checked and only if the job process
is terminated that job should be assumed done. Otherwise, job
reply may be sent for termination when the job is still left
running.

When 'suid' feature is used, setup process is forced so job
processes can be killed by setup process (that runs as suid).