Added support for 'depends' etc. for SharedJobCluster.
Added support for 'depends' and 'setup_args' of NodeAllocate when
used with SharedJobCluster.

Default client port for SharedJobCluster is same that for JobCluster
(61590) instead of 0 (as in last change) so that port forwarding
works. Only when client is used on same computer that runs
dispyscheduler that one needs to set client port to some other
port.