'close_node' and 'resubmit_node' are now async.
'close_node' and 'resubmit_node' may take time so don't let the
caller wait for them to finish. If the caller needs to know the
status of these calls, 'cluster_status' feature can be used to
process state changes. Alternately, these methods could return
the (async) tasks created but until now async features are not
exposed to dispy users so may not be appropriate in these cases.