Moved 'sleep' inside try/except for INT.
When quitting, dispynode raises KeyboardInterrupt within a task,
which is caught in main thread to quit. However, 'time.sleep' and
'send' are outside of try/except of KeyboardInterrupt. This sometimes
causes traceback when SIGINT is received when executing these
statements. To avoid this, these statements are moved inside
try/except block.