Separate ip_addr and ext_ip_addr parameters.
Each 'ext_ip_addr' was initially supposed to be associated with
corresponding 'ip_addr' similar to masquerading for a network.
However, it is possible for a node with a single IP address to
reach nodes in multiple networks if it is behind a router.
To handle such cases in a generic way, ext_ip_addr is now
decoupled from ip_addr.

This needs further work in dispyscheduler and dispynode.