Send pong messages concurrently.
If client sends multiple addresses (with 'ext_ip_addr'), it is
possible some of the addresses are not reachable from node, so
instead of sending reply sequentially (which may take long time
for socket timeout, causing long delays in node's use), use tasks
to send replies concurrently.

In addition to 'ext_ip_addrs', reply is also sent to the address
from which ping message is received.