Increased delay in force termination of job.
When job is cancelled, first send SIGINT/CTRL_BREAK_EVENT to job
(which raises KeyboardInterrupt in the process) and after about
5 seconds, terminate the process. If the job has try/except for
KeyboardInterrupt, then it will have longer time (from 2 seconds
in previous version) to deal with termination. If the job doesn't
deal with KeyboardInterrupt, then it will be terminated right away.