Update __init__.py
The addition of an 'immediate' function call feature to the JobCluster's setup logic allows subsequent items in the 'depends' list to include code that cannot otherwise be defined due to missing module imports at the global level... i.e. declaring a custom class that itself extends a builtin class which must be imported as a base class.